added flow diagram to sidebar
